Refresh all proposals and deleted proposals
Created functionality to refresh all proposals in one go, which updates the
proposal statuses with DELETED if the proposal is no longer existing on the
review side. A proposal with status DELETED is handled as read only, and
disables the ability to add further comments.

There is a known bug that sets the status of all proposals (on the concept proposer side) to
deleted if the server cannot access the concept reviewer's server. Correct
statuses return once a connection can be made.